Added Display of Simulator Information in Graphical View
This commit introduces enhancements to the code that allow the graphical view to display important simulator information in real-time. The modifications include the addition of the simulator running speed, simulation population size, number of sick people, and number of immune people in the graphical view. The display of this information provides users with valuable insights into the simulation's progress and helps them monitor the state of the population more effectively.